requirements.txt
pytest==8.4.2
pytest-xdist==3.6.1
pytest-html==4.1.1
allure-pytest==2.15.0
selenium==4.35.0
webdriver-manager==4.0.2
